requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
blinker==1.4
bookanevents==0.0.5
boto3==1.26.23
botocore==1.29.23
Brlapi==0.8.3
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
commonmark==0.9.1
cryptography==3.4.8
cupshelpers==1.0
cytoolz==0.12.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-storages==1.13.1
djangorestframework==3.14.0
dnspython==2.3.0
docutils==0.19
duplicity==0.8.21
email-validator==1.3.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.10.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fqdn==1.5.1
frozenlist==1.3.1
future==0.18.2
greenlet==2.0.2
hexbytes==0.3.0
httplib2==0.20.2
humanfriendly==10.0
idna==3.4
importlib-metadata==5.1.0
ipfshttpclient==0.8.0a2
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonic==1.0.0
jsonpointer==2.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.11.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lru-dict==1.1.8
macaroonbakery==1.3.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
more-itertools==9.0.0
multiaddr==0.0.9
multidict==6.0.2
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netaddr==0.8.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
panoramix-decompiler==0.4
paramiko==2.9.3
parsimonious==0.8.1
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.9.2
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.5
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-json-logger==2.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
readme-renderer==37.3
regex==2022.10.31
reportlab==3.6.8
requests==2.28.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==12.6.0
rlp==2.0.1
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.1
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.3
sqlparse==0.4.2
stack-data==0.6.2
stevedore==5.0.0
systemd-python==234
terminado==0.17.1
threadpoolctl==3.1.0
timeout-decorator==0.5.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.2.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.12
usb-creator==0.3.7
varint==1.0.2
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
wcwidth==0.2.6
web3==5.31.1
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==9.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==3.11.0
